+++ src/ErrorsInVariables.jl
@@ -1,12 +1,5 @@
 module ErrorsInVariables
 
-using SumTypes
-
-@sum_type Either{A, B} begin
-    Left{A}(::A)
-    Right{B}(::B)
-end
-
 abstract type EiveResult end 
 
 struct SimpleEiveResult <: EiveResult
@@ -23,7 +16,6 @@ include("simex.jl")
 include("iv.jl")
 include("dd.jl")
 
-export Either, Left, Right
 
 export CGA
 export Estimator
@@ -31,7 +23,7 @@ export SimulationExtrapolation
 export IV
 export DD
 
-import .SumTypes: @cases, @sum_type
+
 import .Estimator: eive
 import .Eivem: meive
 import .CGA: cga, cgasample, converged 
@@ -41,7 +33,6 @@ import .SimulationExtrapolation: simex, simex_single_iteration, simex_multiple_i
 import .IV: iv
 import .DD: dd 
 
-export @cases, @sum_type
 
 export eive
 export meive